Add unit tests for the new functions
